Checking to see if Mail.app is running. Installing reindex_mailapp.sh script.

+++ reindex_mailapp.sh
@@ -1,3 +1,12 @@
 #!/bin/sh
 
-rm -f ~/Library/Mail/V2/MailData/Envelope\ Index
+# check to see if Mail.app is running already...
+
+n_mailapp_processes=`ps x | grep -v grep | grep -c "Mail.app"`
+
+if [ "$n_mailapp_processes" -eq "1" ]; then
+  echo "Quit Mail.app first"
+else
+  rm -f ~/Library/Mail/V2/MailData/Envelope\ Index
+  echo "Deleted Envelope Index, restart Mail.app to reindex"
+fi
